Move submessages to own files
