+++ dds/src/implementation/rtps/messages/submessages/ack_nack.rs
@@ -1,0 +1,54 @@
+use crate::implementation::rtps::{
+    messages::{
+        overall_structure::SubmessageHeaderRead, submessage_elements::SequenceNumberSet,
+        types::SubmessageFlag, RtpsMap, SubmessageHeader,
+    },
+    types::{Count, EntityId},
+};
+
+#[derive(Debug, PartialEq, Eq)]
+pub struct AckNackSubmessageRead<'a> {
+    data: &'a [u8],
+}
+
+impl SubmessageHeader for AckNackSubmessageRead<'_> {
+    fn submessage_header(&self) -> SubmessageHeaderRead {
+        SubmessageHeaderRead::new(self.data)
+    }
+}
+
+impl<'a> AckNackSubmessageRead<'a> {
+    pub fn new(data: &'a [u8]) -> Self {
+        Self { data }
+    }
+
+    pub fn final_flag(&self) -> bool {
+        self.submessage_header().flags()[1]
+    }
+
+    pub fn reader_id(&self) -> EntityId {
+        self.map(&self.data[4..])
+    }
+
+    pub fn writer_id(&self) -> EntityId {
+        self.map(&self.data[8..])
+    }
+
+    pub fn reader_sn_state(&self) -> SequenceNumberSet {
+        self.map(&self.data[12..])
+    }
+
+    pub fn count(&self) -> Count {
+        self.map(&self.data[self.data.len() - 4..])
+    }
+}
+
+#[derive(Debug, PartialEq, Eq)]
+pub struct AckNackSubmessageWrite {
+    pub endianness_flag: SubmessageFlag,
+    pub final_flag: SubmessageFlag,
+    pub reader_id: EntityId,
+    pub writer_id: EntityId,
+    pub reader_sn_state: SequenceNumberSet,
+    pub count: Count,
+}
