+++ dds/src/implementation/rtps/messages/submessages/heartbeat.rs
@@ -1,0 +1,62 @@
+use crate::implementation::rtps::{
+    messages::{
+        overall_structure::SubmessageHeaderRead, types::SubmessageFlag, RtpsMap, SubmessageHeader,
+    },
+    types::{Count, EntityId, SequenceNumber},
+};
+
+#[derive(Debug, PartialEq, Eq)]
+pub struct HeartbeatSubmessageRead<'a> {
+    data: &'a [u8],
+}
+
+impl SubmessageHeader for HeartbeatSubmessageRead<'_> {
+    fn submessage_header(&self) -> SubmessageHeaderRead {
+        SubmessageHeaderRead::new(self.data)
+    }
+}
+
+impl<'a> HeartbeatSubmessageRead<'a> {
+    pub fn new(data: &'a [u8]) -> Self {
+        Self { data }
+    }
+
+    pub fn final_flag(&self) -> bool {
+        self.submessage_header().flags()[1]
+    }
+
+    pub fn liveliness_flag(&self) -> bool {
+        self.submessage_header().flags()[2]
+    }
+
+    pub fn reader_id(&self) -> EntityId {
+        self.map(&self.data[4..])
+    }
+
+    pub fn writer_id(&self) -> EntityId {
+        self.map(&self.data[8..])
+    }
+
+    pub fn first_sn(&self) -> SequenceNumber {
+        self.map(&self.data[12..])
+    }
+
+    pub fn last_sn(&self) -> SequenceNumber {
+        self.map(&self.data[20..])
+    }
+
+    pub fn count(&self) -> Count {
+        self.map(&self.data[28..])
+    }
+}
+#[derive(Debug, PartialEq, Eq)]
+pub struct HeartbeatSubmessageWrite {
+    pub endianness_flag: SubmessageFlag,
+    pub final_flag: SubmessageFlag,
+    pub liveliness_flag: SubmessageFlag,
+    pub reader_id: EntityId,
+    pub writer_id: EntityId,
+    pub first_sn: SequenceNumber,
+    pub last_sn: SequenceNumber,
+    pub count: Count,
+}
