+++ dds/src/implementation/rtps/stateful_reader.rs
@@ -19,8 +19,8 @@ use super::{
     messages::{
         overall_structure::RtpsMessageHeader,
         submessages::{
-            DataFragSubmessageRead, DataSubmessageRead, GapSubmessageRead,
-            HeartbeatFragSubmessageRead, HeartbeatSubmessageRead,
+            data::DataSubmessageRead, data_frag::DataFragSubmessageRead, gap::GapSubmessageRead,
+            heartbeat::HeartbeatSubmessageRead, heartbeat_frag::HeartbeatFragSubmessageRead,
         },
     },
     reader::{
