+++ dds/src/implementation/rtps_udp_psm/mapping_rtps_messages/submessages/heart_beat.rs
@@ -1,14 +1,12 @@
-use std::io::{Error, Write};
-
+use super::submessage::MappingWriteSubmessage;
 use crate::implementation::{
     rtps::messages::{
-        overall_structure::SubmessageHeaderWrite, submessages::HeartbeatSubmessageWrite,
+        overall_structure::SubmessageHeaderWrite, submessages::heartbeat::HeartbeatSubmessageWrite,
         types::SubmessageKind,
     },
     rtps_udp_psm::mapping_traits::MappingWriteByteOrdered,
 };
-
-use super::submessage::MappingWriteSubmessage;
+use std::io::{Error, Write};
 
 impl MappingWriteSubmessage for HeartbeatSubmessageWrite {
     fn submessage_header(&self) -> SubmessageHeaderWrite {
