+++ dds/src/implementation/rtps_udp_psm/mapping_rtps_messages/submessages/info_reply.rs
@@ -1,17 +1,13 @@
-use std::io::{Error, Write};
-
-use byteorder::ByteOrder;
-
-use crate::implementation::rtps_udp_psm::mapping_traits::NumberOfBytes;
+use super::submessage::MappingWriteSubmessage;
 use crate::implementation::{
     rtps::messages::{
-        overall_structure::SubmessageHeaderWrite, submessages::InfoReplySubmessageWrite,
-        types::SubmessageKind,
+        overall_structure::SubmessageHeaderWrite,
+        submessages::info_reply::InfoReplySubmessageWrite, types::SubmessageKind,
     },
-    rtps_udp_psm::mapping_traits::MappingWriteByteOrdered,
+    rtps_udp_psm::mapping_traits::{MappingWriteByteOrdered, NumberOfBytes},
 };
-
-use super::submessage::MappingWriteSubmessage;
+use byteorder::ByteOrder;
+use std::io::{Error, Write};
 
 impl MappingWriteSubmessage for InfoReplySubmessageWrite {
     fn submessage_header(&self) -> SubmessageHeaderWrite {
