+++ dds/src/implementation/rtps_udp_psm/mapping_rtps_messages/submessages/info_source.rs
@@ -1,15 +1,13 @@
-use std::io::{Error, Write};
-
-use byteorder::ByteOrder;
-
-use crate::implementation::rtps::messages::submessages::InfoSourceSubmessageWrite;
-use crate::implementation::rtps::messages::{
-    overall_structure::SubmessageHeaderWrite, types::SubmessageKind,
-};
-
-use crate::implementation::rtps_udp_psm::mapping_traits::MappingWriteByteOrdered;
-
 use super::submessage::MappingWriteSubmessage;
+use crate::implementation::{
+    rtps::messages::{
+        overall_structure::SubmessageHeaderWrite,
+        submessages::info_source::InfoSourceSubmessageWrite, types::SubmessageKind,
+    },
+    rtps_udp_psm::mapping_traits::MappingWriteByteOrdered,
+};
+use byteorder::ByteOrder;
+use std::io::{Error, Write};
 
 impl MappingWriteSubmessage for InfoSourceSubmessageWrite {
     fn submessage_header(&self) -> SubmessageHeaderWrite {
@@ -72,27 +70,4 @@ mod tests {
             ]
         );
     }
-
-    // #[test]
-    // fn deserialize_info_source() {
-    //     #[rustfmt::skip]
-    //     let submessage = InfoSourceSubmessageRead::new(&[
-    //         0x0c, 0b_0000_0001, 20, 0, // Submessage header
-    //         0, 0, 0, 0, // unused
-    //         1, 0, 0, 0, //protocol_version | vendor_id
-    //         0, 0, 0, 0, //guid_prefix
-    //         0, 0, 0, 0, //guid_prefix
-    //         0, 0, 0, 0, //guid_prefix
-    //     ]);
-
-    //     let expected_endianness_flag = true;
-    //     let expected_protocol_version = PROTOCOLVERSION_1_0;
-    //     let expected_vendor_id = VENDOR_ID_UNKNOWN;
-    //     let expected_guid_prefix = GUIDPREFIX_UNKNOWN;
-
-    //     assert_eq!(expected_endianness_flag, submessage.endianness_flag());
-    //     assert_eq!(expected_protocol_version, submessage.protocol_version());
-    //     assert_eq!(expected_vendor_id, submessage.vendor_id());
-    //     assert_eq!(expected_guid_prefix, submessage.guid_prefix());
-    // }
 }
